Convert influxdb_metrics sink to use MetricNormalize
This moves the last use of StdMetricNormalize into the influxdb_metrics
sinks, and so also drops all the tests that were specific to that
normalize, as the behavior is covered by the other absolute and
incremental tests.
